countDays()

+++ app/src/main/java/ua/khai/golik/bl/interfaces/CountFinalOrderPrice.java
@@ -8,5 +8,5 @@ public interface CountFinalOrderPrice {
 
     double countPrice(Order order);
 
-    int countDays(Date firstDate, Date secondDate);
+    long countDays(Date firstDate, Date secondDate);
 }
